履歴ページバグ修正・サイドバーフォント変更etc
